standardized to World dimensions

+++ src/main/scala/com/avaglir/knave/map/Landmass.scala
@@ -2,9 +2,9 @@ package com.avaglir.knave.map
 
 import com.avaglir.knave.util._
 
-case class Landmass(sizeClass: IslandClass, tiles: Set[Vector2[Double]], name: Option[String], adjective: Option[String]) {
+case class Landmass(sizeClass: IslandClass, tiles: Set[Vector2[Int]], name: Option[String], adjective: Option[String]) {
   def area = tiles.size
-  lazy val center: Vector2[Double] = tiles.fold(Vector2.ZERO[Double]){ _ + _ } / tiles.size
+  lazy val center: Vector2[Int] = tiles.fold(Vector2.ZERO[Int]){ _ + _ } / tiles.size
   lazy val edge = tiles.filter { tile => tile.adjacent.exists { !tiles.contains(_) }}
 }
